+++ src/components/ui/index.ts
@@ -1,0 +1,8 @@
+export { Button } from "@base-ui/react/button";
+export { Checkbox } from "@base-ui/react/checkbox";
+export { Input } from "@base-ui/react/input";
+export { Radio } from "@base-ui/react/radio";
+export { RadioGroup } from "@base-ui/react/radio-group";
+export { Select } from "@base-ui/react/select";
+export { Switch } from "@base-ui/react/switch";
+export { Tabs } from "@base-ui/react/tabs";